fix loading for install plan
