+++ Client/src/pages/Layout.tsx
@@ -1,5 +1,4 @@
 import {PropsWithChildren, useContext} from "react";
-import IndicatorStripe from "../components/IndicatorStripe";
 import StateContext from "@/providers/StateContext.tsx";
 
 export default function Layout({children}: PropsWithChildren) {
@@ -8,19 +7,18 @@ export default function Layout({children}: PropsWithChildren) {
     return <>
         {!ready &&
             <div
-                className={"z-50 bg-gray opacity-50 absolute top-20 left-0 right-0 bottom-0 flex justify-center font-bold"}>
+                className="z-50 bg-gray opacity-50 absolute top-20 left-0 right-0 bottom-0 flex justify-center font-bold">
                 Please wait...
             </div>
         }
         <div id={"header"}>
             <div className={"frame flex-shrink-0"}>
-                <img src={'/svg/logo-full.svg'}/>
+                <img src={'/svg/logo-full.svg'} alt={'Frierun'}/>
             </div>
             <div className={"flex-shrink-0 pt-2 xxl:ml-[15%]"}>
-                <img src={'/svg/deco.svg'}/>
+                <img src={'/svg/deco.svg'} alt={'Decoration'}/>
             </div>
         </div>
-        <IndicatorStripe running={!ready}/>
         <div className={"frame my-6"}>
             {children}
         </div>
